# openapi.yml
openapi: 3.0.3
info:
  title: Rarus Test API
  version: 1.0.0
servers:
  - url: /api
paths:
  /categories:
    get:
      summary: Get active categories with product counts
      responses:
        '200':
          description: OK
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                parent_id: { type: string, format: uuid, nullable: true }
                is_active: { type: boolean }
              required: [name]
      responses:
        '201': { description: Created }
  /categories/{id}:
    put:
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                parent_id: { type: string, format: uuid, nullable: true }
                is_active: { type: boolean }
      responses:
        '200': { description: OK }
    delete:
      summary: Soft delete category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /categories/tree:
    get:
      summary: Get grouped products by categories tree
      responses:
        '200': { description: OK }
  /products:
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                category_id: { type: string, format: uuid }
                price: { type: number }
                is_active: { type: boolean }
              required: [name, category_id]
      responses:
        '201': { description: Created }
  /products/{id}:
    put:
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                category_id: { type: string, format: uuid }
                price: { type: number }
                is_active: { type: boolean }
      responses:
        '200': { description: OK }
    delete:
      summary: Soft delete product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /products/by-category/{categoryId}:
    get:
      summary: Get active products by category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }